🚚 Move `interface` to `pages`
