+++ food_classifier/src/service_ui/pages/customer_page.py
